requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
PyMuPDF
python-docx
python-pptx
pandas
chromadb
